CRUD for trips Api complete
