+++ wonderlust/src/app/api/trips/route.ts
@@ -32,123 +32,3 @@ export async function GET(req: NextRequest) {
   );
 }
 }
-
-export async function POST(req: NextRequest) { 
- 
- try {
-  await connectMongo();
-  const body = await req.json()
-
-  const { tripname, budget, travellers,
-   startDay, endDay,
-   destination ,userId} = body;
-  const parsedStartDay = new Date(startDay);
-  const parsedEndDay = new Date(endDay);
-
-  if (!tripname || !budget || !travellers || !startDay || !endDay || !destination || !userId) {
-    return NextResponse.json(
-      { success: false, message: "Missing required fields" },
-      { status: HttpStatusCode.BadRequest }
-    );
-  }
-  
-  const newTrip = new Trip({
-   tripname:tripname,
-   budget:budget,
-   travellers:travellers,
-   destination: destination,
-   startDay:parsedStartDay,
-   endDay: parsedEndDay,
-   user:new mongoose.Types.ObjectId(userId)
-  });
-
-   await Trip.create(newTrip)
-  
-     return NextResponse.json({ success: true, newTrip }, { status: HttpStatusCode.Created, });
-     
- } catch (error) {
-  console.error("POST Error:", error);
-  return NextResponse.json(
-    { success: false, message: "Server Error", error: error.message },
-    { status: HttpStatusCode.InternalServerError }
-  );
-}
-}
-
-export async function DELETE(req: NextRequest) {
-  try {
-    await connectMongo();
-    const { tripId } = await req.json();
-
-    const tripCheck = await Trip.findOne({_id: new mongoose.Types.ObjectId(tripId), user: new mongoose.Types.ObjectId(tempUserID)})
-    
-    if (tripCheck) {
-       await Trip.findByIdAndDelete(tripId);
-
-    return NextResponse.json(
-      { success: true, message: "Trip deleted successfully"},
-      { status: HttpStatusCode.Ok }
-    );
-    }
-    return NextResponse.json(
-      { success: false, message: "Authentication failed or missing fields or trip not found" },
-      { status: HttpStatusCode.InternalServerError }
-    );
-  }
-  catch (error) {
-    console.error("Delete Error:", error);
-    return NextResponse.json(
-      { success: false, message: "Server Error", error: error.message },
-      { status: HttpStatusCode.InternalServerError }
-    );
-  }
-}
-
-export async function PUT(req: NextRequest) {
-  try {
-    await connectMongo();
-    const body = await req.json()
-  
-    const { tripId ,tripname, budget, travellers,
-     startDay, endDay,
-     destination ,userId} = body;
-  
-    if (!tripname || !budget || !travellers || !startDay || !endDay || !destination || !userId) {
-      return NextResponse.json(
-        { success: false, message: "Missing required fields" },
-        { status: HttpStatusCode.BadRequest }
-      );
-    }
-
-    const updatedData = 
-      {
-        tripname:tripname,
-        budget:budget,
-        travellers:travellers,
-        destination: destination,
-        startDay: startDay,
-        endDay: endDay,
-        user:new mongoose.Types.ObjectId(userId)
-       }
-    const checker = await Trip.findByIdAndUpdate(tripId, updatedData)
-    if (checker) {
-      return NextResponse.json(
-        { success: true, message: "Trip updated successfully"},
-        { status: HttpStatusCode.Ok }
-      );
-    }
-   
-    return NextResponse.json(
-        { success: false, message: "Trip not found"},
-        { status: HttpStatusCode.BadRequest }
-      );
-    
-  }
-  catch (error) {
-    console.error("Delete Error:", error);
-    return NextResponse.json(
-      { success: false, message: "Server Error", error: error.message },
-      { status: HttpStatusCode.InternalServerError }
-    );
-  }
-}
